+++ ikuzo/storage/x/elasticsearch/mapping/v2.go
@@ -66,7 +66,10 @@ var v2Mapping = `{
 						"docType": {"type": "keyword"},
 						"namedGraphURI": {"type": "keyword"},
 						"entryURI": {"type": "keyword"},
-						"modified": {"type": "date"}
+						"modified": {"type": "date"},
+						"sourceID": {"type": "keyword"},
+						"sourcePath": {"type": "keyword"},
+						"groupID": {"type": "keyword"}
 					}
 				},
 				"protobuf": {
@@ -194,6 +197,14 @@ func V2MappingUpdate() string {
 // 'strict' on dynamic creating of new fields in the index.
 var v2MappingUpdate = `{
   "properties": {
+	"meta": {
+		"type": "object",
+		"properties": {
+			"sourceID": {"type": "keyword"},
+			"sourcePath": {"type": "keyword"},
+			"groupID": {"type": "keyword"}
+		}
+	},
     "tree": {
       "properties": {
         "physDesc": {"type": "keyword"},