+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ### Added 
 
+- Initial service for elasticsearch functionality. Better mapping management.  [[GH-78]](https://github.com/delving/hub3/pull/78)
 - Multitenancy support through organisationIDs  [[GH-77]](https://github.com/delving/hub3/pull/77)
 - Initial version of Time Revision Store  [[GH-76]](https://github.com/delving/hub3/pull/76)
 - Dedicated SPARQL service that can be used as a separate publisher [[GH-75]](https://github.com/delving/hub3/pull/75)